refactor: make use of module js to split popup.js into multiple files
