+++ manifest.json
@@ -1,6 +1,5 @@
 {
 	"manifest_version": 2,
-
 	"name": "Moodle Downloader",
 	"description": "A chrome extension for downloading Moodle resources",
 	"version": "1.5.3",
