+++ src/popup.html
@@ -3,10 +3,13 @@
 
 <head>
   <meta charset="UTF-8">
+  <!-- METADATA -->
   <title>Moodle Downloader</title>
-  <script src="./scripts/tailwind.min.js"></script>
-  <script src="./scripts/jszip.min.js"></script>
-  <script src="./popup.js"></script>
+  <!-- LIBRARIES -->
+  <script src="./libraries/tailwind.min.js"></script>
+  <script src="./libraries/jszip.min.js"></script>
+  <!-- SCRIPTS -->
+  <script type="module" src="./popup.js"></script>
 </head>
 
 <body>
